refactor : product 도메인 분리
