+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/product/exception/ProductExceptionMessage.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.exception;
+package com.sparta.blackwhitedeliverydriver.product.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
