+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/product/service/ProductService.java
@@ -1,23 +1,21 @@
-package com.sparta.blackwhitedeliverydriver.service;
+package com.sparta.blackwhitedeliverydriver.product.service;
 
-import com.sparta.blackwhitedeliverydriver.dto.CreateProductRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.ProductIdResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.ProductRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.ProductResponseDto;
-import com.sparta.blackwhitedeliverydriver.entity.Product;
-import com.sparta.blackwhitedeliverydriver.entity.Store;
-import com.sparta.blackwhitedeliverydriver.entity.User;
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import com.sparta.blackwhitedeliverydriver.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.ProductExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.StoreExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.repository.ProductRepository;
-import com.sparta.blackwhitedeliverydriver.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.repository.UserRepository;
-import com.sparta.blackwhitedeliverydriver.security.UserDetailsImpl;
+import com.sparta.blackwhitedeliverydriver.category.dto.CreateProductRequestDto;
+import com.sparta.blackwhitedeliverydriver.product.dto.ProductIdResponseDto;
+import com.sparta.blackwhitedeliverydriver.product.dto.ProductRequestDto;
+import com.sparta.blackwhitedeliverydriver.product.dto.ProductResponseDto;
+import com.sparta.blackwhitedeliverydriver.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.product.exception.ProductExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.store.exception.StoreExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.product.repository.ProductRepository;
+import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.common.security.UserDetailsImpl;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
